requirements.txt
numpy
pandas
matplotlib
notebook
seaborn
pyperclip
tqdm
pygame
python-dotenv
torch
torchvision
chess
huggingface_hub[torch]
transformers
bitsandbytes>=0.39.0
tokenizers
sentencepiece
datasets
hf-xet
peft
accelerate
deepspeed
wandb